# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "AgeCategory": {
        "enum": [
          "child",
          "teen",
          "adult",
          "older_adult"
        ],
        "type": "string"
      },
...
      "ArtStyle": {
        "enum": [
          "flat_vector",
          "cartoon",
          "graphic_novel",
          "paper_cutout",
          "silhouette"
        ],
        "type": "string"
      },
...
      "BodyBuild": {
        "enum": [
          "slender",
          "average",
          "sturdy",
          "broad"
        ],
        "type": "string"
      },
...
      "BottomClothing": {
        "enum": [
          "trousers",
          "shorts",
          "skirt"
        ],
        "type": "string"
      },
...
      "CharacterBuildRead": {
        "additionalProperties": false,
        "properties": {
          "character": {
            "$ref": "#/components/schemas/CharacterDefinition"
          },
          "constraints": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/GeneratedConstraintDefinition"
            },
            "title": "Constraints",
            "type": "array"
          },
          "diagnostics": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BuilderDiagnostic"
            },
            "title": "Diagnostics",
            "type": "array"
          },
          "generator_version": {
            "default": "procedural-human-v1",
            "title": "Generator Version",
            "type": "string"
          },
          "normalized_request": {
            "$ref": "#/components/schemas/CharacterBuilderRequest"
          }
        },
        "required": [
          "normalized_request",
          "character"
        ],
        "title": "CharacterBuildRead",
        "type": "object"
      },
...
      "CharacterBuilderRequest": {
        "additionalProperties": false,
        "properties": {
          "age_category": {
            "$ref": "#/components/schemas/AgeCategory",
            "default": "adult"
          },
          "bottom": {
            "$ref": "#/components/schemas/BottomClothing",
            "default": "trousers"
          },
          "build": {
            "$ref": "#/components/schemas/BodyBuild",
            "default": "average"
          },
          "face_shape": {
            "$ref": "#/components/schemas/FaceShape",
            "default": "oval"
          },
          "footwear": {
            "$ref": "#/components/schemas/Footwear",
            "default": "shoes"
          },
          "hair_style": {
            "$ref": "#/components/schemas/HairStyle",
            "default": "short"
          },
          "height": {
            "$ref": "#/components/schemas/HeightClass",
            "default": "average"
          },
          "name": {
            "default": "Procedural Human",
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "outerwear": {
            "$ref": "#/components/schemas/Outerwear",
            "default": "none"
          },
          "palette": {
            "$ref": "#/components/schemas/CharacterPalette",
            "default": {
              "accent": "#f0d36b",
              "bottom": "#2f3a56",
              "hair": "#3d2a1e",
              "shoes": "#2b2b2b",
              "skin": "#c98f62",
              "top": "#2f6f73"
            }
          },
          "presentation": {
            "$ref": "#/components/schemas/Presentation",
            "default": "neutral"
          },
          "proportions": {
            "$ref": "#/components/schemas/CharacterProportions",
            "default": {
              "arm_length": 1.0,
              "asymmetry": 0.0,
              "head_size": 1.0,
              "hip_width": 1.0,
              "leg_length": 1.0,
              "shoulder_width": 1.0,
              "torso_length": 1.0,
              "waist_width": 1.0
            }
          },
          "style": {
            "$ref": "#/components/schemas/ArtStyle",
            "default": "flat_vector"
          },
          "top": {
            "$ref": "#/components/schemas/TopClothing",
            "default": "tshirt"
          }
        },
        "title": "CharacterBuilderRequest",
        "type": "object"
      },
      "CharacterDefinition": {
        "additionalProperties": false,
        "properties": {
          "attachments": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AttachmentDefinition"
            },
            "title": "Attachments",
            "type": "array"
          },
          "id": {
            "pattern": "^char_[a-z0-9][a-z0-9_-]*$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "rig": {
            "$ref": "#/components/schemas/RigDefinition"
          }
        },
        "required": [
          "id",
          "name",
          "rig"
        ],
        "title": "CharacterDefinition",
        "type": "object"
      },
      "CharacterGenerationRequest": {
        "description": "AI character generation request (Ollama planner -> deterministic builder).",
        "properties": {
          "description": {
            "default": "",
            "maxLength": 4000,
            "title": "Description",
            "type": "string"
          },
          "expected_revision": {
            "title": "Expected Revision",
            "type": "string"
          },
          "form": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CharacterBuilderRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "model": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "temperature": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          }
        },
        "required": [
          "model",
          "expected_revision"
        ],
        "title": "CharacterGenerationRequest",
        "type": "object"
      },
      "CharacterMutationResult": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/ProjectDocument"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          }
        },
        "required": [
          "document",
          "revision"
        ],
        "title": "CharacterMutationResult",
        "type": "object"
      },
      "CharacterPalette": {
        "additionalProperties": false,
        "properties": {
          "accent": {
            "default": "#f0d36b",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Accent",
            "type": "string"
          },
          "bottom": {
            "default": "#2f3a56",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Bottom",
            "type": "string"
          },
          "hair": {
            "default": "#3d2a1e",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Hair",
            "type": "string"
          },
          "shoes": {
            "default": "#2b2b2b",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Shoes",
            "type": "string"
          },
          "skin": {
            "default": "#c98f62",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Skin",
            "type": "string"
          },
          "top": {
            "default": "#2f6f73",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Top",
            "type": "string"
          }
        },
        "title": "CharacterPalette",
        "type": "object"
      },
      "CharacterProportions": {
        "additionalProperties": false,
        "properties": {
          "arm_length": {
            "default": 1.0,
            "title": "Arm Length",
            "type": "number"
          },
          "asymmetry": {
            "default": 0.0,
            "title": "Asymmetry",
            "type": "number"
          },
          "head_size": {
            "default": 1.0,
            "title": "Head Size",
            "type": "number"
          },
          "hip_width": {
            "default": 1.0,
            "title": "Hip Width",
            "type": "number"
          },
          "leg_length": {
            "default": 1.0,
            "title": "Leg Length",
            "type": "number"
          },
          "shoulder_width": {
            "default": 1.0,
            "title": "Shoulder Width",
            "type": "number"
          },
          "torso_length": {
            "default": 1.0,
            "title": "Torso Length",
            "type": "number"
          },
          "waist_width": {
            "default": 1.0,
            "title": "Waist Width",
            "type": "number"
          }
        },
        "title": "CharacterProportions",
        "type": "object"
      },
      "CharacterRead": {
        "properties": {
          "character": {
            "$ref": "#/components/schemas/CharacterDefinition"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "character",
          "revision"
        ],
        "title": "CharacterRead",
        "type": "object"
      },
      "CharacterWrite": {
        "properties": {
          "character": {
            "$ref": "#/components/schemas/CharacterDefinition"
          },
          "expected_revision": {
            "title": "Expected Revision",
            "type": "string"
          }
        },
        "required": [
          "character",
          "expected_revision"
        ],
        "title": "CharacterWrite",
        "type": "object"
      },
...
      "FaceShape": {
        "enum": [
          "round",
          "oval",
          "square",
          "heart",
          "long"
        ],
        "type": "string"
      },
...
      "Footwear": {
        "enum": [
          "shoes",
          "boots",
          "sneakers"
        ],
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HairStyle": {
        "enum": [
          "bald",
          "short",
          "bob",
          "curly",
          "long",
          "coily"
        ],
        "type": "string"
      },
...
      "HeightClass": {
        "enum": [
          "short",
          "average",
          "tall"
        ],
        "type": "string"
      },
...
      "Job": {
        "description": "Public, serializable snapshot of a job.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_detail": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Detail"
          },
          "error_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Kind"
          },
          "id": {
            "pattern": "^job_[a-z0-9][a-z0-9_-]*$",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "progress": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/JobProgress"
            },
            "title": "Progress",
            "type": "array"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JsonValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "state",
          "created_at",
          "updated_at"
        ],
        "title": "Job",
        "type": "object"
      },
...
      "JobState": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "type": "string"
      },
...
      "Outerwear": {
        "enum": [
          "none",
          "vest",
          "coat"
        ],
        "type": "string"
      },
...
      "Presentation": {
        "enum": [
          "masculine",
          "feminine",
          "neutral"
        ],
        "type": "string"
      },
...
      "TopClothing": {
        "enum": [
          "tshirt",
          "shirt",
          "sweater",
          "jacket"
        ],
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "RigStory Studio",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/characters/build": {
      "post": {
        "operationId": "build_character_api_v1_characters_build_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterBuilderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterBuildRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Character",
        "tags": [
          "characters"
        ]
      }
    },
...
    "/api/v1/characters/{character_id}": {
      "delete": {
        "operationId": "delete_character_api_v1_characters__character_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "title": "Character Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_revision",
            "required": true,
            "schema": {
              "title": "Expected Revision",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterMutationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Character",
        "tags": [
          "characters"
        ]
      },
      "get": {
        "operationId": "get_character_api_v1_characters__character_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "title": "Character Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Character",
        "tags": [
          "characters"
        ]
      },
      "patch": {
        "operationId": "update_character_api_v1_characters__character_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "title": "Character Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterMutationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Character",
        "tags": [
          "characters"
        ]
      }
    },
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_api_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/events": {
      "get": {
        "operationId": "job_events_api_v1_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Events",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/characters": {
      "get": {
        "operationId": "list_project_characters_api_v1_projects__project_id__characters_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CharacterDefinition"
                  },
                  "title": "Response List Project Characters Api V1 Projects  Project Id  Characters Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Project Characters",
        "tags": [
          "characters"
        ]
      },
      "post": {
        "operationId": "create_project_character_api_v1_projects__project_id__characters_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterMutationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project Character",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/v1/projects/{project_id}/characters/generate": {
      "post": {
        "description": "Start an AI character-generation job (specs \u00a723.2, returns 202 + a job).",
        "operationId": "generate_project_character_api_v1_projects__project_id__characters_generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Project Character",
        "tags": [
          "characters"
        ]
      }
    },
...
  }
}